Tweaks
The padding between the edges of the screen and the main container was
off compared to the my courses page. I removed the course-unit-wrapper
from the div for the column and into the article tag to correct this.

I also added the option for a video description to be entered into the
unit template.